+++ src/TextRenderer.cs
@@ -2,12 +2,7 @@ using OpenTK.Graphics.OpenGL4;
 using OpenTK.Mathematics;
 using System;
 using System.Collections.Generic;
-
-namespace EOCS
-
-{
-    
-}
+namespace EOCS;
 
 public class SlowTextRenderer
 {