+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 # All changes in project EOCS
 
+## [0.0.4] - 2026.04.17
+
+### Added
+- Generation Text Atlas
+- Based text class (GlyphData)
+
 ## [0.0.3] - 2026.04.17
 
 ### Changed